VRT 足りなかったテスト追加 不要なテストの削除
