+++ app/components/sections/Faq/Faq.stories.tsx
@@ -1,0 +1,16 @@
+import type { Meta, StoryObj } from "@storybook/nextjs-vite";
+import Faq from "./index";
+
+const meta: Meta<typeof Faq> = {
+  title: "Sections/Faq",
+  component: Faq,
+  tags: ["autodocs"],
+  parameters: {
+    layout: "fullscreen",
+  },
+};
+
+export default meta;
+type Story = StoryObj<typeof Faq>;
+
+export const Default: Story = {};
